I'm now using a 'model' factory for every page, that is responsible for arranging all of the data for a specific controller, so that it fits the view. In fact, it builds the real view-model.
Such 'model' services are allowed to use one another in order to mix data of different entities. They are also a way to share data between pages (i.e. classModel holds a list of students, while studentModel uses that service in order to find the specific student it needs to work with).