1. I split the 'Add class' process into 2 steps: first u associate the class with a course. Once the course is chosen, you go to the 2nd step where you edit the rest of the class' settings. The course cannot be changed once set. That helps us by avoiding the logic of changing a course for an existing class, that might already have lots of data associated with the course (like the permissions for every student per bp in the course).
2. The navigation between 'classes list' -> 'single class' -> 'single student' is now changed.
You have a 'save' button that stays in the same page, and a 'back' button which goes back.
When you're saving, you're saving the whole class.
This way the hierarchy between 'class' and 'student' is simpler - you can't really edit and save a single student, without saving the whole class, and now the view itself supports this notion and clarifies it for the user.