As I introduce the pages for the student, I did some refactoring of the hierarchy of pages/directories/states - which all correspond to one another.
The main pages are now of 'trainer', 'student' and 'login' - this conforms to the actual 1st level of separation in the application's pages. 'trainer' page will have only the trainer pages under it, and so on.

Then there are directories to match these pages, and angular states to match them. Everything is organized according to this hierarchy.