RAV-4020 Prevent duplicate username
Changing the implementation - instead of having a unique index on the username field in the DB, I now perform the validation of username uniqueness in the application level (server side). The reason for this is that the same user can be in more than one class. With the current architecture of the application, such a case will result in 2 different student entities, each associated with 2 different user entities, that will both have the same username.

 So, the username uniqueness in fact exists only within a class.
  That's why the validation should be application-level, and done only on the users of the same class.