RAV-4006  add text search in the "select blueprint for course"
Added the functionality of the search in a more visible way.
Up until now, the search text area was part of the DataGrid component (a built-in feature), but had to be opened with a little arrow at the top-right.
Now it's a visible search text area at the top of the popup.