Fixing scale problems:
1. When fetching all of the Apps of a single Class, we'd make many GET /applications/appId calls to Ravello, which is very inefficient.
Now we make a single GET /applications, and filter the ones relevant to the Class in code.

2. When creating a new App in a Class, there was a bug.
The problem was that the UI initiated a single, separate POST /rest/applications against the Training Server, for every App that's to create.
But since the document that's being updated is the Class, updating it with many such async calls caused data inconsistencies.
I changed it, so that creating many Apps together, is done in a single POST /rest/applications, which updates the Class in one stroke.