Adding a feature that creates Apps (by the Trainer) in chunks.
Technically this is crucial because we encountered several limitations in creating many Apps altogether:
- A timeout problem in the Training Portal server.
- Rate limits of the Ravello Management server against the various cloud providers.

In terms of UX, this means that the Create action taken by the Trainer will return instantly.
In accompany, a auto-refresh mechanism will fetch the up-to-date status of the Apps.
(cherry picked from commit fecd5a4)