Adding a feature to edit the publish details of the Apps that will be created during a Class. The feature enables to edit this information only on the Class level - i.e. for each BP that's in the Class (via its Course), the Trainer can determine the publish method/cloud/region and auto-stop.
This required some changes in how the relationship between Courses, Classes and Apps is made.
Basically, it's now the classes-service that has the responsibility to combine the Class with its Course model.
It then works directly with the courses-service to do so.
The cache (of either) is intimately altered when the API function of the service are called (getting Apps, for instance).
(cherry picked from commit f49c31b)