Error with getting VNC.
Problem was with a new version of the superagent 3rd-party library.
Might be that the real problem is that the current Ravello API demands a Content-Type header of application/json, while the result is not JSON...

In addition, I changed the error message returned to the user, to be more properly displayed (even though not much more informative, since this is a server error).