Fix bug of the browser infinitely trying to access the first, base URL of the application.
After login, the application attempts to navigate to the first state, but if it fails, it will keep trying to navigate to this state (and fail).
This was a problem in the urlRouter and its redirection mechanism.
If we control the state navigation ourselves (using $state instead of $urlRouterProvider), failures of navigation won't create those infinite loops.