Improve error handling and reporting to the client.
Now there are safe-nets that catch errors in any case, and are tailored to deal with superagent (i.e. MGMT) or Mongo (DB) errors.
We eventually produce a readable string to the client, and logging all information we have for future analysis.

This also improves the readablity of the code, as now there are no longer many error-handling code pieces in the controllers.
They are split between the services, DAL, and the generic error handler.