Change the Create Apps flow for Trainer.
We have two things in mind:
1) The UX currently is confusing and bug-prone, since there's no good indication of the various actions taken against Ravello in the background, so the user doesn't know what happens - mainly which applications are created.
2) Doing the whole flow against Ravello (created -> publish -> autostop) before updating the Class and the DB, is more likely to cause a lack of sync between Ravello and the TP.
For instance, an App might be created, but not published - it will exist in Ravello, but not in the TP.

The new flow that addresses both issues:

1) The Trainer selects Apps and clicks Create.
2) Request is sent to the TP.
3) All of the Create REST calls are sent to MGMT together.
4) When they are all done (whether successfully or not), the DB of the TP is updated (only with the successful ones).
5) Response is returned to the client.
6) Now the Trainer sees all of the created Apps in the UI, and the DB is properly synced with Ravello.
7) In the background, asynchronously, the created Apps are published and their autostop is set - in chunks, to avoid time-based limits set on the various clouds.
8) UI is auto-refreshed, and will be updated with the new status of the Apps, as time goes by...

Possible pitfall - there are rate limits defined on MGMT itself, to prevent flooding with REST calls per organization.
Now, when we're doing all of the Create calls against MGMT together, we might reach this rate limit.
This is a known limitation of the Ravello product, and if a Trainer runs into it, one should ask for changing that limit for the orgnaization in Ravello...