Bug fix - when changing the Blueprints of a Course, existing Classes of that Course were not affected correctly.
Now, I only half-solve the problem in the server - Classes that reference the Course are not changed, so that they won't reference Blueprints that were removed from the Course.
But for new Blueprints that were added to the Course, and do not yet exist in the Classes - they are NOT added in the server side.
The reason is that currently, the default values to set for these new entities in the Class, are determined in the client-side code.
I don't want to replicate them to server-side.
We need a solution in which all such properties are determined only in server-side, and are sent to the client.

But, the problem is still solved in the client-side, that knows to match the newly-added Blueprints of the Course, with the current data of the Class, and display it correctly.