Bug fix - course option is not set when adding a new class.
In the latest revision, we had a fix to a bug in the relations between Class and Course.
Due to the fix, we had a regression in the flow of adding a new Class - the requirement is to have the first Course selected by default, in the first page of selecting the Course of the Class. Not having any selection is just annoying...