Bug fix - batch delete.
We had a theory that there might be a bug in deleting Apps of the same Class, in separate REST calls, since they eventually all look on the very same document in the DB (the Class), and dur to race conditions, one REST call might fetch the Class and attempt to update it, while in the meantime another call had already changed that Class document.
That's why the way to go for us is to combine all Apps deletions to a single REST call, that also make only a single Class update.

We theorized this after we saw that on a client's TP, some students had more than one App (of the very same BP, which shouldn't happen).
We saw that such a corruption occurred when more than one App was deleted (with several REST calls being made).
We then came up with this theory.

Unfortunately, I'm very short on time, and reproducing race conditions is tough - so I didn't even attempt to prove our theory.
I just rewrote the code, to work in batch.

 @benbracha - I only made basic tests against our dev babu, which seemed fine.
 Also, I didn't upgrade the VMs we prepare in our RND account.
 I just thought that I better give this effort now, before I go on the vacation, just in case it proves important to some clients.