RAV-7498: Training Portal - Ravello password sent in the open
Not sending Ravello in HTTP when not essential.
In the communication between TP client and server, the Ravello password should not be passed, unless it's an essential part of the communication (i.e. when an Admin user creates a Trainer user, and determines its Ravello credentials).
This is a way to reduce to the necessary minimum the times this password is revealed.