Adding a script that accepts a CSV of students, and sets them in a class. This way, it's easier to configure many students, and set different BP permissions for each.
The CSV format:
FirstName,surname,username,password,startVmsPermission,stopVmsPermission,restartVmsPermission,consolePermission

Limitations:
1) This script overrides all existing students (and their data) in the class.
If a student has an app associated with it in the class, running this script will break the association.
The app will remain running in Ravello, but will not be associated to the Training Portal and the student anymore.
2) One can set the BP permissions for every student independently, but for all BPs of the class altogether.