Bug fix - ravello apps not deleted from students data in the DB.
After a trainer deletes students apps, the cache that holds the class data in the browser remains intact, i.e. still holds those apps, even though the DB is updated.
If the trainer would then go and update the class, the data of the student apps (that's still in the browser's cache) would find its way to the DB.